+++ backend/src/main/java/toonpick/app/auth/handler/OAuth2SuccessHandler.java
@@ -3,6 +3,7 @@ package toonpick.app.auth.handler;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
@@ -11,23 +12,21 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationSu
 import org.springframework.stereotype.Component;
 import toonpick.app.auth.user.CustomOAuth2UserDetails;
 import toonpick.app.auth.jwt.JwtTokenProvider;
-import toonpick.app.auth.service.AuthService;
+import toonpick.app.auth.service.TokenService;
+import toonpick.app.service.MemberService;
 
 import java.io.IOException;
 import java.util.Collection;
 
 @Component
+@RequiredArgsConstructor
 public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
 
     private final JwtTokenProvider jwtTokenProvider;
-    private final AuthService authService;
+    private final TokenService tokenService;
+    private final MemberService memberService;
 
     private static final Logger logger = LoggerFactory.getLogger(OAuth2SuccessHandler.class);
-
-    public OAuth2SuccessHandler(JwtTokenProvider jwtTokenProvider, AuthService authService) {
-        this.authService = authService;
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -37,10 +36,10 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
         String role = authorities.stream().findFirst().map(GrantedAuthority::getAuthority).orElse("");
 
-        Long userid = authService.getUserIdByUsername(username);
+        Long userid = memberService.getUserIdByUsername(username);
         String refreshToken = jwtTokenProvider.createRefreshToken(userid, username, role);
 
-        authService.saveRefreshToken(username, refreshToken);
+        tokenService.saveRefreshToken(username, refreshToken);
 
         response.addCookie(jwtTokenProvider.createCookie("refresh", refreshToken));
