+++ backend/src/main/java/toonpick/app/service/MemberService.java
@@ -3,6 +3,7 @@ package toonpick.app.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.dto.MemberDTO;
 import toonpick.app.entity.Member;
 import toonpick.app.mapper.MemberMapper;
@@ -20,9 +21,17 @@ public class MemberService {
         this.memberMapper = memberMapper;
     }
 
+    @Transactional(readOnly = true)
     public MemberDTO getUserById(Long id) {
         Member member = memberRepository.findById(id)
                                   .orElseThrow(() -> new UsernameNotFoundException("Member not found"));
         return memberMapper.userToUserDto(member);
+    }
+
+    @Transactional(readOnly = true)
+    public Long getUserIdByUsername(String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("Member not found with username: " + username));
+        return member.getId();
     }
 }
