+++ backend/src/test/java/toonpick/app/integration/TokenReissueIntegrationTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 import toonpick.app.auth.jwt.JwtTokenProvider;
 import toonpick.app.auth.jwt.JwtTokenValidator;
-import toonpick.app.auth.service.AuthService;
+import toonpick.app.auth.service.TokenService;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
@@ -32,7 +32,7 @@ class TokenReissueIntegrationTest {
     private JwtTokenProvider jwtTokenProvider;
 
     @MockBean
-    private AuthService authService;
+    private TokenService tokenService;
 
     @MockBean
     private JwtTokenValidator jwtTokenValidator;
@@ -45,7 +45,7 @@ class TokenReissueIntegrationTest {
 
         when(jwtTokenValidator.extractRefreshTokenFromCookies(any())).thenReturn(refreshToken);
         doNothing().when(jwtTokenValidator).validateRefreshToken(refreshToken);
-        when(authService.refreshAccessToken(refreshToken)).thenReturn(newAccessToken);
+        when(tokenService.renewAccessToken(refreshToken)).thenReturn(newAccessToken);
 
         mockMvc.perform(post("/api/reissue")
                 .cookie(new Cookie("refresh", refreshToken)))
